LayoutTests/html5lib/resources/comments01.dat

#data
FOO<!-- BAR -->BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR  -->
|     "BAZ"

#data
FOO<!-- BAR --!>BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR  -->
|     "BAZ"

#data
FOO<!-- BAR --   >BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR --    -->
|     "BAZ"

#data
FOO<!-- BAR -- <QUX> -- MUX -->BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX  -->
|     "BAZ"

#data
FOO<!-- BAR -- <QUX> -- MUX --!>BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX  -->
|     "BAZ"

#data
FOO<!-- BAR -- <QUX> -- MUX -- >BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX --  -->
|     "BAZ"

#data
FOO<!---->BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"

#data
FOO<!--->BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"

#data
FOO<!-->BAZ
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"
